#19 [feat] Memo 컴포넌트 분리
